WIP: keep header visible during login
