+++ packages/demo/src/components/App/index.js
@@ -3,13 +3,11 @@ import React from 'react'
 import {connect} from 'cerebral/react'
 import translations from '../../common/computed/translations'
 
-import LangSelector from '../LangSelector'
 import Login from '../Login'
 import Menu from '../Menu'
 import Timer from '../Timer'
 import Workspace from '../Workspace'
-
-const TaglineRe = /^(.*)\[Cerebral\](.*)$/
+import Header from '../Header'
 
 export default connect(
   {
@@ -22,56 +20,20 @@ export default connect(
     signOutClicked: 'user.signOutClicked'
   },
   function Demo ({t, loggedIn, loading, currentUser, signOutClicked}) {
-    const match = TaglineRe.exec(t.SiteTagLine)
-    const tagline = match
-      ? <h2 className='subtitle'>{match[1]}<a href='http://cerebraljs.com'>Cerebral</a>{match[2]}</h2>
-      : <h2 className='subtitle'>{t.SiteTagLine}</h2>
     if (loading) {
       return null
     }
     if (!loggedIn) {
-      return <Login />
+      return (
+        <div className='App'>
+          <Header />
+          <Login />
+        </div>
+      )
     }
     return (
       <div className='App'>
-        <section className='hero is-dark'>
-          <div className='hero-head'>
-            <div className='container'>
-              <nav className='nav'>
-                <div className='nav-left'>
-                  <a className='nav-item is-brand'>
-                    {/* Wait until we have a logo that scales well in this place
-                      <img src='/cerebral-logo.png' alt='Cerebral logo' />
-                    */}
-                  </a>
-                </div>
-                <div className='nav-right'>
-                  {currentUser && (
-                    <div className='nav-item'>
-                      {currentUser.email}
-                    </div>
-                  )}
-                  {currentUser && (
-                    <div className='nav-item'>
-                      <a href='#' onClick={() => signOutClicked()}>
-                        {t.loginSignOut}
-                      </a>
-                    </div>
-                  )}
-                  <div className='nav-item'>
-                    <LangSelector />
-                  </div>
-                </div>
-              </nav>
-            </div>
-          </div>
-          <div className='hero-body'>
-            <div className='container'>
-              <h1 className='title'>{t.CerebralDemo}</h1>
-              {tagline}
-            </div>
-          </div>
-        </section>
+        <Header />
         <section className='section'>
           <div className='container'>
             <div className='columns'>
